fix(sqlite): simplify sidecar permission hardening
